+++ src/CLI/Program.cs
@@ -1,4 +1,4 @@
-﻿using AMS2CM;
+﻿using Core;
 
 try
 {
