+++ src/Core/Core.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0-windows</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <Company>OpenSimTools</Company>
+        <Authors>OpenSimTools Contributors</Authors>
+        <LangVersion>10</LangVersion>
+        <AssemblyName>AMS2CM.Core</AssemblyName>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Gameloop.Vdf" Version="0.6.2" />
+        <PackageReference Include="Gameloop.Vdf.JsonConverter" Version="0.2.1" />
+        <PackageReference Include="SevenZipExtractor" Version="1.0.17" />
+    </ItemGroup>
+</Project>
+
