+++ src/Core/IMod.cs
@@ -1,0 +1,10 @@
+﻿namespace Core;
+
+public interface IMod
+{
+    string PackageName { get; }
+    bool Installed { get; }
+    IReadOnlyCollection<string> InstalledFiles { get; }
+
+    void Install(string dstPath);
+}
