+++ src/Core/IPostProcessor.cs
@@ -1,0 +1,6 @@
+﻿namespace Core;
+
+public interface IPostProcessor
+{
+    void PerformPostProcessing(List<IMod> filesByMod);
+}
